refactor(handler): extract DoReadAll from ReadAllWeb
